# docs/swagger.yaml
basePath: /federation/v1
definitions:
...
  dto.GetArtefactResponse:
    properties:
      appProviderId:
        type: string
      artefactDescription:
        type: string
      artefactDescriptorFileFormat:
        $ref: '#/definitions/models.ArtefactFileFormat'
      artefactDescriptorType:
        $ref: '#/definitions/models.ArtefactDescriptorType'
      artefactId:
        type: string
      artefactName:
        type: string
      artefactVersionInfo:
        type: string
      artefactVirtType:
        $ref: '#/definitions/models.ArtefactVirtType'
      objectRepoLocation:
        $ref: '#/definitions/models.ObjectRepoLocation'
    type: object
...
  models.ArtefactDescriptorType:
    enum:
    - HELM
    - TERRAFORM
    - ANSIBLE
    - SHELL
    - COMPONENTSPEC
    type: string
    x-enum-varnames:
    - HELM
    - TERRAFORM
    - ANSIBLE
    - SHELL
    - COMPONENTSPEC
  models.ArtefactFileFormat:
    enum:
    - ZIP
    - TAR
    - TEXT
    - TARGZ
    type: string
    x-enum-varnames:
    - ZIP
    - TAR
    - TEXT
    - TARGZ
  models.ArtefactVirtType:
    enum:
    - VM_TYPE
    - CONTAINER_TYPE
    type: string
    x-enum-varnames:
    - VM_TYPE
    - CONTAINER_TYPE
...
  models.InvalidParam:
    properties:
      param:
        type: string
      reason:
        type: string
    type: object
...
  models.ObjectRepoLocation:
    properties:
      password:
        description: Password to access the repository
        type: string
      repoURL:
        type: string
      token:
        description: Authorization token to access the repository
        type: string
      userName:
        description: Username to access the repository
        type: string
    type: object
...
  models.ProblemDetails:
    properties:
      cause:
        description: Fixed string indicating cause of the issue
        type: string
      detail:
        description: Specific detail of the issue
        type: string
      invalidParams:
        items:
          $ref: '#/definitions/models.InvalidParam'
        type: array
      title:
        description: Summary of the problem
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is the API documentation for the MEC Federator.
  title: MEC Federator API
  version: "1.0"
paths:
...
  /ewbi/{federationContextId}/artefact:
    post:
      consumes:
      - multipart/form-data
      description: Receives and onboards an artefact from the origin operator. The
        artefact is a packaged file (zip/tar) containing application scripts and packaging
        files. The artefact is validated and stored in the database.
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Artefact package file (zip/tar format)
        in: formData
        name: artefactFile
        required: true
        type: file
      - description: Unique identifier for the artefact
        in: formData
        name: artefactId
        required: true
        type: string
      - description: Application provider identifier
        in: formData
        name: appProviderId
        required: true
        type: string
      - description: Human-readable name of the artefact
        in: formData
        name: artefactName
        required: true
        type: string
      - description: Version information of the artefact
        in: formData
        name: artefactVersionInfo
        required: true
        type: string
      - description: Detailed description of the artefact
        in: formData
        name: artefactDescription
        required: true
        type: string
      - description: Virtualization type (e.g., container, VM)
        in: formData
        name: artefactVirtType
        required: true
        type: string
      - description: Descriptor file type (e.g., TOSCA, Helm)
        in: formData
        name: artefactDescriptorType
        required: true
        type: string
      - description: File format of the artefact (e.g., zip, tar.gz)
        in: formData
        name: artefactFileFormat
        required: true
        type: string
      - description: Original filename of the artefact
        in: formData
        name: artefactFileName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: Artefact onboarded successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid form data or validation errors
          schema:
            $ref: '#/definitions/models.ProblemDetails'
        "500":
          description: File processing or database errors
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Onboard Artefact
      tags:
      - EWBI - ArtefactManagement
  /ewbi/{federationContextId}/artefact/{artefactId}:
    delete:
      consumes:
      - application/json
      description: Removes an artefact from the database. Artefacts that have been
        onboarded to the orchestrator cannot be deleted until the onboarding is removed.
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Artefact identifier
        in: path
        name: artefactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: Artefact deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Artefact has been onboarded and cannot be deleted
          schema:
            $ref: '#/definitions/models.ProblemDetails'
        "404":
          description: Artefact not found
          schema:
            $ref: '#/definitions/models.ProblemDetails'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Delete Artefact
      tags:
      - EWBI - ArtefactManagement
  /ewbi/{federationContextId}/artefacts/{artefactId}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific artefact by its
        ID within a federation context
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Artefact identifier
        in: path
        name: artefactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetArtefactResponse'
        "404":
          description: Artefact not found
          schema:
            $ref: '#/definitions/models.ProblemDetails'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Get Artefact Details
      tags:
      - EWBI - ArtefactManagement
...
swagger: "2.0"